server/requirements.txt
Django>=4.2.0
djangorestframework
django-cors-headers
psycopg2-binary
pyjwt
drf-yasg
requests
